+++ src/main.rs
@@ -13,18 +13,29 @@ fn main(boot_info: &'static BootInfo) -> ! {
     moros::init(boot_info);
     print!("\x1b[?25h"); // Enable cursor
     loop {
-        let bootrc = "/ini/boot.sh";
-        if sys::fs::File::open(bootrc).is_some() {
-            usr::shell::main(&["shell", bootrc]);
+        if let Some(cmd) = option_env!("MOROS_CMD") {
+            let prompt = usr::shell::prompt_string(true);
+            println!("{}{}", prompt, cmd);
+            usr::shell::exec(cmd);
+            sys::acpi::shutdown();
         } else {
-            if sys::fs::is_mounted() {
-                println!("Could not find '{}'", bootrc);
-            } else {
-                println!("MFS is not mounted to '/'");
-            }
-            println!("Running console in diskless mode");
-            usr::shell::main(&["shell"]);
+            user_boot();
         }
+    }
+}
+
+fn user_boot() {
+    let script = "/ini/boot.sh";
+    if sys::fs::File::open(script).is_some() {
+        usr::shell::main(&["shell", script]);
+    } else {
+        if sys::fs::is_mounted() {
+            println!("Could not find '{}'", script);
+        } else {
+            println!("MFS is not mounted to '/'");
+        }
+        println!("Running console in diskless mode");
+        usr::shell::main(&["shell"]);
     }
 }